requirements.txt
psycopg2-binary
apache-airflow[fab]